Extract component methods from fluxMixin into separate module
